requirements.txt
PyGithub
selenium
PyMuPDF
webdriver-manager
python-dotenv
